requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via -r requirements.in
cryptography==41.0.7
    # via jwt
distlib==0.3.8
    # via virtualenv
filelock==3.13.1
    # via virtualenv
google-api-core==2.15.0
    # via google-api-python-client
google-api-python-client==2.111.0
    # via -r requirements.in
google-auth==2.26.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
googleapis-common-protos==1.62.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
jwt==1.3.1
    # via -r requirements.in
nodeenv==1.8.0
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
platformdirs==4.1.0
    # via virtualenv
pre-commit==3.6.0
    # via -r requirements.in
protobuf==4.25.1
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pyparsing==3.1.1
    # via httplib2
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.1.0
    # via requests
virtualenv==20.25.0
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
